#3370 move missing files to stimmzettel directory
