+++ wls-gui-wahllokalsystem/src/components/dse/stimmzettelerfassung/TheEingabehistorieCard.vue
