+++ wls-gui-wahllokalsystem/stories/components/dse/stimmzettelerfassung/TheEingabehistorieCard.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/vue3-vite";
 
-import TheEingabehistorieCard from "@/components/dse/TheEingabehistorieCard.vue";
+import TheEingabehistorieCard from "@/components/dse/stimmzettelerfassung/TheEingabehistorieCard.vue";
 import { InputHistoryTypeEnum } from "@/types/dse/stimmzettelerfassung/InputHistoryTypeEnum.ts";
 
 const meta = {
