+++ wls-gui-wahllokalsystem/stories/components/dse/stimmzettelerfassung/TheStimmzettelCommandProcessingTextField.stories.ts
@@ -4,7 +4,7 @@ import type { Meta, StoryObj } from "@storybook/vue3-vite";
 
 import { useWahlvorschlaegeTestDataFactory } from "@tests/utils/wahlvorschlaege/WahlvorschlaegeTestDataFactory.ts";
 
-import TheStimmzettelCommandProcessingTextField from "@/components/dse/TheStimmzettelCommandProcessingTextField.vue";
+import TheStimmzettelCommandProcessingTextField from "@/components/dse/stimmzettelerfassung/TheStimmzettelCommandProcessingTextField.vue";
 import { useStimmzettelManager } from "@/composables/dse/stimmzettelerfassung/stimmzettelManager.ts";
 
 const meta = {
